% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationAnalysis.R
\name{mutationPercentage}
\alias{mutationPercentage}
\title{Get count and percentage of gene's affected by mutation}
\usage{
mutationPercentage(mutationData)
}
\arguments{
\item{mutationData}{A data frame with column's icgc_donor_id,
and gene affected. The data frame should be in the same format as the icgc
database's DCC Data Release.}
}
\value{
Returns a data frame with results
\itemize{
\item gene_affected - Gene code for designated mutated gene in donors
\item mutation_count - The number of mutation's observed in sample for
gene_affected
\item percentage_of_sample_with_mutated_gene - The percentage of the
frequency of observed mutated gene in sample
}
}
\description{
A function that analyzes the provided data by getting the
percentage of each observed mutation's frequency in the sample.
The data frame must contain column's icgc_donor_id,
and gene_affected.
}
\examples{
# Example 1:
# Using icgc_data available inside package

# Get gene's mutation frequency, and percentage in sample
mutationPercentageResults <- mutationPercentage(icgc_data)
mutationPercentageResults

}
\references{
Hadley Wickham, Romain Francois, Lionel Henry and Kirill Muller
(2022). dplyr: A Grammar of Data Manipulation. R package version
1.0.10. https://CRAN.R-project.org/package=dplyr
}
