% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationAnalysis.R
\name{mutationTypePlot}
\alias{mutationTypePlot}
\title{Bar plot and pie chart for mutation type data samples}
\usage{
mutationTypePlot(mutationData)
}
\arguments{
\item{mutationData}{A data frame with column's icgc_mutation_id,
cds_mutation, and consequence_type. The data frame should be in the same
format as the icgc database's DCC data release}
}
\value{
Returns a list of two plots.
\itemize{
\item mutationTypes - A pie chart showing the count and mutation types
undergone in the sample
\item substitutionTypes - A bar plot showing the count and substitution
types undergone in the sample
}
}
\description{
A function that returns a list containing a list of plots for mutation type
data samples
}
\examples{
mutationTypePlotResults <- mutationTypePlot(icgc_data)
mutationTypePlotResults["mutationTypes"]
mutationTypePlotResults["substitutionTypes"]

}
