% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rundiagMutAnalysis.R
\name{rundiagMutAnalysis}
\alias{rundiagMutAnalysis}
\title{Launch Shiny App for diagMutAnalysis}
\usage{
rundiagMutAnalysis()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for diagMutAnalysis
The purpose of this app is only to illustrate how a Shiny
app works. The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

diagMutAnalysis::rundiagMutAnalysis()
}

}
\references{
Jung, J. (2022) diagMutAnalysis: R package that analyzes somatic
mutation data for BCB410H: Applied Bioinformatics. Unpublished.
URL https://github.com/wjdwogud24/diagMutAnalysis.
}
