% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icgc_data}
\alias{icgc_data}
\title{Data description}
\format{
An object of class \code{data.frame} with 4776 rows and 11 columns.
}
\usage{
icgc_data
}
\description{
Data description
}
\keyword{datasets}
