% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationAnalysis.R
\name{mutationTypePlot}
\alias{mutationTypePlot}
\title{Bar plot and pie chart for mutation type data samples}
\usage{
mutationTypePlot(mutationData, ch = "A", population = nrow(mutationData))
}
\arguments{
\item{mutationData}{A data frame with columns icgc_mutation_id,
cds_mutation, and consequence_type. The data frame should be in the same
format as the icgc databases DCC data release}

\item{ch}{A character value indicating which chromosome
wants to be viewed. Default value is A which stands for all chromosomes. Can
choose one specific chromosome to view plots with.}

\item{population}{An integer indicating the size of the sample wanting to use
for the plot. Non randomized population extraction. Takes sample from 1 to
population number. If a non integer value is taken will round it
using the round() function. Resulting plot will not equal population value
since it is using population number of samples.
Not population number of results.}
}
\value{
Returns a list of two plots.
\itemize{
\item mutation_types - A pie chart showing the count and mutation types
undergone in the sample
\item substitution_types - A bar plot showing the count and substitution
types undergone in the sample
}
}
\description{
A function that returns a list of two plots a bar plot and pie chart
for mutation type analysis
}
\examples{
# Example 1:
mutationTypePlotResults <- mutationTypePlot(icgc_data)
#To display each plot
mutationTypePlotResults["mutation_types"]
mutationTypePlotResults["substitution_types"]

#Example 2:
mutationTypePlotResults <- mutationTypePlot(icgc_data, ch = "X", population = 100)
#To display each plot
mutationTypePlotResults["mutation_types"]
mutationTypePlotResults["substitution_types"]

}
\references{
Hadley Wickham, Romain Francois, Lionel Henry and Kirill Muller
(2022). dplyr: A Grammar of Data Manipulation. R package version
1.0.10. https://CRAN.R-project.org/package=dplyr

H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New
York, 2016.

Hadley Wickham (2022). stringr: Simple, Consistent Wrappers for Common
String Operations. R package version 1.4.1.
https://CRAN.R-project.org/package=stringr
}
