% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rundiagMutAnalysis.R
\name{rundiagMutAnalysis}
\alias{rundiagMutAnalysis}
\title{Launch Shiny App for diagMutAnalysis}
\usage{
rundiagMutAnalysis()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for diagMutAnalysis
The purpose of this app is only to illustrate how a Shiny
app works. The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

diagMutAnalysis::rundiagMutAnalysis()
}

}
\references{
Silva, A. (2022) TestingPackage: An Example R Package For
BCB410H. Unpublished. https://github.com/anjalisilva/TestingPackage
}
